+++ tournament/tables/src/main/java/com/dss/tennis/tournament/tables/model/dto/AbstractSequentialDTO.java
@@ -8,5 +8,5 @@ import lombok.*;
 @AllArgsConstructor
 @NoArgsConstructor
 public abstract class AbstractSequentialDTO {
-    private Integer sequenceNumber;
+    private Byte sequenceNumber;
 }
